# argox-project/argox-collector/openapi.json
{
  "components": {
    "schemas": {
      "ApiKeyCreate": {
        "description": "Request body for ``POST /api/v1/keys``.",
        "properties": {
          "expires_in": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "name",
          "scopes"
        ],
        "title": "ApiKeyCreate",
        "type": "object"
      },
      "ApiKeyCreateResponse": {
        "description": "Create response: the metadata view plus the one-time raw secret.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "scopes",
          "created_at",
          "revoked",
          "key"
        ],
        "title": "ApiKeyCreateResponse",
        "type": "object"
      },
      "ApiKeyListResponse": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyView"
            },
            "title": "Keys",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "keys",
          "total"
        ],
        "title": "ApiKeyListResponse",
        "type": "object"
      },
      "ApiKeyView": {
        "description": "Non-secret view of a stored key.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "scopes",
          "created_at",
          "revoked"
        ],
        "title": "ApiKeyView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TraceDetailResponse": {
        "description": "Full span waterfall returned by ``GET /api/v1/traces/{trace_id}``.\n\n``truncated`` is True when the trace exceeded the index's per-trace span\nceiling and ``spans`` was cut to keep the response bounded.",
        "properties": {
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanDetail"
            },
            "title": "Spans",
            "type": "array"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "trace_id",
          "spans"
        ],
        "title": "TraceDetailResponse",
        "type": "object"
      },
      "TraceListResponse": {
        "description": "Paginated payload returned by ``GET /api/v1/traces``.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TraceSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "TraceListResponse",
        "type": "object"
      },
      "TraceSummary": {
        "description": "Aggregated, per-trace row for the dashboard list view.",
        "properties": {
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "span_count": {
            "title": "Span Count",
            "type": "integer"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "span_count"
        ],
        "title": "TraceSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Server-side ingestion, indexing and policy distribution service for the Argox observability platform.",
    "title": "Argox Collector",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/keys": {
      "get": {
        "description": "List every key's non-secret metadata, newest first.",
        "operationId": "list_keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List API keys",
        "tags": [
          "keys"
        ]
      },
      "post": {
        "description": "Mint a new API key and return its raw secret once.",
        "operationId": "create_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an API key",
        "tags": [
          "keys"
        ]
      }
    },
    "/api/v1/keys/{key_id}": {
      "delete": {
        "description": "Revoke a key. Idempotent: revoking an unknown/already-revoked key is 404.",
        "operationId": "revoke_key",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key revoked (no body)."
          },
          "404": {
            "description": "No active key with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke an API key",
        "tags": [
          "keys"
        ]
      }
    },
...
    "/api/v1/traces": {
      "get": {
        "description": "List trace summaries, newest first.",
        "operationId": "list_traces",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List traces",
        "tags": [
          "query"
        ]
      }
    },
    "/api/v1/traces/{trace_id}": {
      "get": {
        "description": "Return the full span waterfall of one trace.",
        "operationId": "get_trace",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get trace detail",
        "tags": [
          "query"
        ]
      }
    },
...
    "/v1/traces": {
      "post": {
        "description": "Accept an OTLP ``ExportTraceServiceRequest`` and persist its spans.\n\nValidates the batch synchronously, then delegates the blob write and index\ninsert to a background task and returns ``202 Accepted``. When the request\ncarries ``X-Argox-Durable: true`` the persistence runs synchronously and the\nendpoint returns ``200 OK`` only once the data is committed.",
        "operationId": "ingest_traces",
        "responses": {
          "200": {
            "content": {
              "application/json": {},
              "application/x-protobuf": {}
            },
            "description": "OTLP ExportTraceServiceResponse. Serialized as protobuf or JSON to match the request Content-Type."
          },
          "415": {
            "description": "Unsupported Content-Type."
          }
        },
        "summary": "OTLP/HTTP trace ingest",
        "tags": [
          "traces"
        ]
      }
    }
  }
}